[MOVE/#355] 설정 폴더 분리
