+++ app/src/main/java/com/el/yello/presentation/main/profile/info/ProfileFragment.kt
@@ -16,7 +16,7 @@ import com.el.yello.R
 import com.el.yello.databinding.FragmentProfileBinding
 import com.el.yello.presentation.main.profile.ProfileViewModel
 import com.el.yello.presentation.main.profile.detail.ProfileDetailActivity
-import com.el.yello.presentation.main.profile.manage.ProfileManageActivity
+import com.el.yello.presentation.setting.SettingActivity
 import com.el.yello.presentation.pay.PayActivity
 import com.el.yello.util.Utils.setPullToScrollColor
 import com.el.yello.util.amplitude.AmplitudeUtils
@@ -24,7 +24,6 @@ import com.el.yello.util.context.yelloSnackbar
 import com.example.domain.entity.ProfileUserModel
 import com.example.ui.activity.navigateTo
 import com.example.ui.base.BindingFragment
-import com.example.ui.fragment.toast
 import com.example.ui.view.UiState
 import com.example.ui.view.setOnSingleClickListener
 import dagger.hilt.android.AndroidEntryPoint
@@ -88,7 +87,7 @@ class ProfileFragment : BindingFragment<FragmentProfileBinding>(R.layout.fragmen
     private fun initProfileManageBtnListener() {
         binding.btnProfileManage.setOnSingleClickListener {
             AmplitudeUtils.trackEventWithProperties("click_profile_manage")
-            activity?.navigateTo<ProfileManageActivity>()
+            activity?.navigateTo<SettingActivity>()
             viewModel.resetStateVariable()
         }
     }
