+++ app/src/main/java/com/el/yello/presentation/setting/ProfileQuitDialog.kt
@@ -1,4 +1,4 @@
-package com.el.yello.presentation.main.profile.manage
+package com.el.yello.presentation.setting
 
 import android.content.Context
 import android.content.Intent
