+++ app/src/main/java/com/el/yello/presentation/setting/ProfileQuitOneActivity.kt
@@ -1,9 +1,9 @@
-package com.el.yello.presentation.main.profile.manage
+package com.el.yello.presentation.setting
 
 import android.content.Intent
 import android.os.Bundle
 import com.el.yello.R
-import com.el.yello.databinding.ActivityProfileQuitTwoBinding
+import com.el.yello.databinding.ActivityProfileQuitOneBinding
 import com.el.yello.util.amplitude.AmplitudeUtils
 import com.example.ui.base.BindingActivity
 import com.example.ui.view.setOnSingleClickListener
@@ -11,26 +11,32 @@ import dagger.hilt.android.AndroidEntryPoint
 import org.json.JSONObject
 
 @AndroidEntryPoint
-class ProfileQuitTwoActivity :
-    BindingActivity<ActivityProfileQuitTwoBinding>(R.layout.activity_profile_quit_two) {
+class ProfileQuitOneActivity :
+    BindingActivity<ActivityProfileQuitOneBinding>(R.layout.activity_profile_quit_one) {
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
         initBackBtnListener()
+        initReturnBtnListener()
         initQuitBtnListener()
     }
 
     private fun initBackBtnListener() {
-        binding.btnProfileQuitBack.setOnSingleClickListener { finish() }
+        binding.btnProfileQuitForSureBack.setOnSingleClickListener { finish() }
+    }
+
+    private fun initReturnBtnListener() {
+        binding.btnProfileQuitReturn.setOnSingleClickListener { finish() }
     }
 
     private fun initQuitBtnListener() {
-        binding.btnProfileQuitForSure.setOnSingleClickListener {
+        binding.btnProfileQuitResume.setOnSingleClickListener {
             AmplitudeUtils.trackEventWithProperties(
                 "click_profile_withdrawal",
-                JSONObject().put("withdrawal_button", "withdrawal3"),
+                JSONObject().put("withdrawal_button", "withdrawal2"),
             )
-            Intent(this, ProfileQuitReasonActivity::class.java).apply {
+            Intent(this, ProfileQuitTwoActivity::class.java).apply {
                 addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
                 startActivity(this)
             }
