+++ app/src/main/java/com/el/yello/presentation/setting/ProfileQuitReasonActivity.kt
@@ -1,4 +1,4 @@
-package com.el.yello.presentation.main.profile.manage
+package com.el.yello.presentation.setting
 
 import android.os.Bundle
 import androidx.activity.viewModels
